+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 ### Deprecated
 - \Fiv\Form\Element\Base use \Fiv\Form\Element\BaseElement
 - \Fiv\Form\Element\BaseElement::required
+- \Fiv\Form\Element\Validator\Len
 - \Fiv\Form\Validator\Base::i
 - \Fiv\Form\Validator\Base use \Fiv\Form\Validator\BaseValidator
 - \Fiv\Form\Filter\Base::i